requirements.txt

kedro-viz
kedro-datasets
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter